Integrated input prompts for creating, reading, updating, and deleting Appointment, and Doctor records and created a Runner class for the operations to keep the main class clean.
